The left-hand-side index is now accurate
Still need to figure out how to handle the case where the buffer is
either larger or smaller than a single line. Right now, those cases
produce garbage in the output.

Now the program opens an input file in read-only mode.

There is an integer overflow if the number of columns to print is odd! I
need to fix that.