C include style dump now uses the hex LUT
This output style is ~22 times faster than the "standard" xxd.

Fixed a bug in the c include style dump that would incorrectly provide
some details that xxd does not when an explicit variable name is not
given.

Fixed a bug in the c include style dump where xxd-zig would leave a
trailing comma. The "standard" xxd does not do this.