Cleaned up the code for autoskip
Replaced my custom slice-testing code with the more-capable function in
the zig standard library.

Fixed the bug that caused the c include style output to print a variable
name when the user did not ask for it. The behavior now is closer to the
"standard" xxd.

Removed a TODO note about using a buffered reader since we are already
using a buffered reader.