C include style capitalisation works
I re-wrote it to get rid of the need to use a text buffer. This version
is perhaps a bit slower, but it's much easier to understand and speed
shouldn't matter for something that is printed twice per dump.

Changed the code to pass around a pointer to the print parameters
instead of copying the struct instance at every function call.

Made the command-line arguments const since the program shouldn't be
changing them.

Added a couple TODO notes for later.

Shuffled a couple things around in main() so that stuff matched the
command-line agument order (where appropriate).