Bug fixes and some simplifications, one new test
Tests were broken due to a code change that required print parameters to
be passed as a pointer instead of copied to functions.

Simplified the color logic in the print_columns() function using a
switch statement instead of a nest of ifs and elses.

Added a test for the ASCII output function.