Initial commit of working data to bin string conversion
Initial benchmarks show ~3x speed improvement

Added a few tests to validate the binary string LUT. The table is built
at compile-time and not written out in the source code, so a couple
checks for correctness seemed important.

All tests pass

Corrected bug where input filenames for the c-include mode would not get
automatically used as the name. This was a deviation from the original
xxd's behavior.

Added some more detail to the readme file.