Added better error handling, fixed bugs
Fixed a bug where certain sequences in the data would confuse the
reverse parsing system.

Added somewhat graceful error handling in case there is a problem with
reverse mode like the user forgetting the -p or -a options.

Started adding the scaffolding for even better error handling.

Currently the performance of the reverse mode is bad: 50 % worse than xxd