Improvements to usability
Added usage notes to the readme

Converted a message about reversing little-endian and C include style
dumps into an error to match the existing code style.

Added some helpful error messages for the user in case they try to do
something that is not supported or is nonsensical.

Eliminated a test that served no purpose.