Lots of improvements to this buffered input version
Input strings now work again!

Fixed memory leaks in the dump and reverse dump functions.

Tests for reverse mode are working again and passing.

Reverse mode errors are once again caught and the user is told what
went wrong.

Added a new error handler for cases where the input is over a memory
page long without a newline - this is probably not a hexdump.

"Plain" or postscript dumps are as-of-yet untested.

Still some ugly code to fix...