Fixed reversing "plain" postscript dumps
Also fixed a bug in "plain" postscript dumps that cropped up for files
larger than a page. Moving to buffered inputs caused this. Fixing this
required updating how plain dumps are printed, which also meant changing
some of the unit tests.

C-include style dumps are also broken for inputs larger than a page. A
fix is in-progress as of this commit.