Fixed seek and length options
Introducting a manual buffer for reads caused a lot of problems with
the seek and length features for dumps. This commit fixes them.

Added note to add a warning when using the seek and length options in
reverse mode. The original xxd ignores these options and I am fine with
that behavior for this program. However, the user should be warned about
this!