Minor optimizations, better error handling
Slightly sped up normal dumps (~1% improvement) by simplifying a
function call and a specific write that happens all the time.

Added a message that gets sent when you try to seek or shorten the
output of a reverse dump. The original xxd just ignores the seek and
shortening options, while this program at least warns the user that
these options are ignored in reverse mode.

Added a test that validates the seek and length options.