Wrapping up this feature
In the end, smaller files are little bit slower with the buffered input
vs how things work in the main branch. However, small files take only a
couple milliseconds, so this shouldn't be a big deal.

Larger files are still much faster than xxd, so that's good.

This commit brings a couple of extremly minor optimisations that I found
while using the time profiler in Xcode's Instruments.

Some variables were renamed so that the code's intent is more clear.

A bug with "plain" postscript dumps was fixed.

Readme updated with details that reflect this branch.